fix(wattpm): properly auto resolve nested package manager
