+++ packages/wattpm/test/build.test.js
@@ -71,7 +71,7 @@ test('install - should install dependencies of application and its services usin
   const installProcess = await wattpm('install', rootDir, '-P', 'pnpm')
 
   ok(installProcess.stdout.includes('Installing dependencies for the application using pnpm ...'))
-  ok(installProcess.stdout.includes('Installing dependencies for the service main using pnpm ...'))
+  ok(installProcess.stdout.includes('Installing dependencies for the service main using npm ...'))
 })
 
 test('install - should respect the service package manager, if any', async t => {
@@ -106,5 +106,5 @@ test('install - should install production dependencies only', async t => {
   const installProcess = await wattpm('install', rootDir, '-p', '-P', 'pnpm')
 
   ok(installProcess.stdout.includes('Installing production dependencies for the application using pnpm ...'))
-  ok(installProcess.stdout.includes('Installing production dependencies for the service main using pnpm ...'))
+  ok(installProcess.stdout.includes('Installing production dependencies for the service main using npm ...'))
 })
